+++ app/components/shared/compact-menu.hbs
@@ -13,7 +13,7 @@
     <AuIcon @icon="user-fill" @alignment="left" />
     Mandatarissen
   </AuLink>
-  <AuLink @route="verkiezingen" role="menuitem">
+  <AuLink @route="verkiezingen.installatievergadering" role="menuitem">
     <AuIcon @icon="vote" @alignment="left" />
     Verkiezingen
   </AuLink>
