+++ app/components/shared/main-menu.hbs
@@ -16,7 +16,7 @@
       </AuNavigationLink>
     </li>
     <li class="au-c-list-navigation__item">
-      <AuNavigationLink @route="verkiezingen">
+      <AuNavigationLink @route="verkiezingen.installatievergadering">
         <AuIcon @icon="vote" @alignment="left" />
         Verkiezingen
       </AuNavigationLink>
