+++ app/router.js
@@ -49,13 +49,8 @@ Router.map(function () {
   });
 
   this.route('verkiezingen', function () {
-    this.route(
-      'verkiezingsuitslag',
-      { path: '/:bestuursorgaan_in_tijd_id/verkiezingsuitslag' },
-      function () {
-        this.route('prepare');
-      }
-    );
+    this.route('verkiezingsuitslag');
+    this.route('installatievergadering');
   });
 
   this.route('leidinggevenden', function () {
