+++ app/templates/index.hbs
@@ -46,7 +46,10 @@
           <:title>Verkiezingen</:title>
           <:description>Hier vind je een overzicht van de verkiezingen.</:description>
           <:link>
-            <AuLink @route="verkiezingen" @skin="button">Ga naar verkiezingen</AuLink>
+            <AuLink
+              @route="verkiezingen.installatievergadering"
+              @skin="button"
+            >Ga naar verkiezingen</AuLink>
           </:link>
         </LoketModuleCard>
       </li>
